Removed ValuesController; finished refactor
The backend-api refactor should be completed and we should be on a
properly-constructed Web API backend.  Corrected a minor bug with the
query string path in downloading a file.
