+++ WebFrontend/Controllers/ServerInfoController.cs
@@ -1,7 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Net;
 using System.Net.Http;
+using System.Net.Http.Headers;
 using System.Web.Http;
 using WebFrontend.Models;
 
@@ -31,7 +33,6 @@ namespace WebFrontend.Controllers
                                   };
         }
 
-        
         // PUT api/serverinfo
         public HttpResponseMessage Put(HttpRequestMessage request, [FromBody]ConnectionInfo connection)
         {
